+++ geotrek/core/views.py
@@ -511,7 +511,7 @@ class TrailFormatList(MapEntityFormat, TrailList):
         "districts",
         "areas",
         "uuid",
-        "decoupled",
+        "coupled",
         *AltimetryMixin.COLUMNS,
     ]