+++ geotrek/signage/templates/signage/sql/post_10_views.sql
@@ -25,7 +25,7 @@ CREATE VIEW {{ schema_geotrek }}.v_signages AS WITH v_signage_tmp AS
             t.type_id,
             CONCAT (e.min_elevation, 'm') AS elevation,
             e.geom,
-            e.decoupled
+            e.coupled
         FROM signage_signage t
             left join core_topology e on t.topo_object_id = e.id
             left join signage_signagetype b on t.type_id = b.id
@@ -70,7 +70,7 @@ SELECT a.id,
        {% endfor %}
        a.elevation AS "Elevation",
        a.publication_date AS "Insertion date",
-       a.decoupled AS "Decoupled geometry",
+       a.coupled AS "Network-coupled",
        a.geom
 FROM v_signage_tmp a
 LEFT JOIN signage_signagetype b ON a.type_id = b.id